requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
openai>=1.3.0
tiktoken>=0.5.1
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.0
pandas>=2.0.0
